Restrured loading of the Ontology.
The OntologyMetricManagerImpl class is no longer needed. The application
now uses the out of the box tools of the OWLAPI. Thus, we safe lines of
code and additonally get rid of the need to store the ontology file
temporarly locally. Now, the ontology always stays in the application as
an object.